Move category form to separate file
