+++ frontend/sumo-championship/src/pages/AddSeason.tsx
@@ -1,24 +1,17 @@
-import React, { useRef, useState } from 'react';
+import React, { useState } from 'react';
 import './../styles/Pages.css';
 import Button from '../components/Atoms/Button';
 import TextField from '../components/molecules/TextField';
 import { useNavigate } from 'react-router-dom';
 import ROUTES from '../routes/ROUTES';
-import Checkbox from '../components/Atoms/Checkbox';
-import MinMaxField from '../components/molecules/MinMaxField';
-import CategoryTable from '../components/organisms/CategoryTable';
-import { Category, Gender } from '../types/Category';
-import { set } from 'zod';
+import { Category } from '../types/Category';
+import CategoryForm from '../components/organisms/CategoryForm';
 
 const errorPointsValues = {
-  categoryName: 2,
-  gender: 3,
-  age: 5,
-  weight: 7,
-};
-
-const checkIfNameIsUnique = (array: Category[], name: string) => {
-  return !array.some((category) => category.name === name);
+  name: 2,
+  startDate: 3,
+  endDate: 5,
+  startEndRange: 7,
 };
 
 const AddSeason: React.FC = () => {
@@ -26,135 +19,17 @@ const AddSeason: React.FC = () => {
   const [name, setName] = useState('');
   const [startDate, setStartDate] = useState('');
   const [endDate, setEndDate] = useState('');
-  const [categoryName, setCategoryName] = useState('');
-  const [minAge, setMinAge] = useState(16);
-  const [maxAge, setMaxAge] = useState(100);
-  const [minWeight, setMinWeight] = useState(40);
-  const [maxWeight, setMaxWeight] = useState(200);
   const [categories, setCategories] = useState<Category[]>([]);
-  const [errorPoints, setErrorPoints] = useState(1);
-  const [editedCategoryNumber, setEditedCategoryNumber] = useState(-1);
-  const isEdited = editedCategoryNumber > -1;
 
-  const femaleCheckboxRef = useRef<HTMLInputElement>(null);
-  const maleCheckboxRef = useRef<HTMLInputElement>(null);
+  const [errorPoints, setErrorPoints] = useState(1);
 
   const addSeason = () => {
-    console.log('Add season');
-  };
-
-  const resetCategoryForm = () => {
-    setCategoryName('');
-    setMinAge(16);
-    setMaxAge(100);
-    setMinWeight(40);
-    setMaxWeight(200);
-    setErrorPoints(1);
-    setEditedCategoryNumber(-1);
-  };
-
-  const checkCategoryFormErrors = () => {
-    let points = 1;
-    if (!categoryName) {
-      points *= errorPointsValues.categoryName;
-    }
-
-    if (minAge > maxAge) {
-      points *= errorPointsValues.age;
-    }
-
-    if (minWeight > maxWeight) {
-      points *= errorPointsValues.weight;
-    }
-
-    setErrorPoints(points);
-    return points;
-  };
-
-  const getGender = () => {
-    let gender: Gender = 'All';
-    if (
-      !!femaleCheckboxRef.current?.checked &&
-      !!maleCheckboxRef.current?.checked
-    ) {
-      gender = 'All';
-    } else if (femaleCheckboxRef.current?.checked) {
-      gender = 'Female';
-    } else if (maleCheckboxRef.current?.checked) {
-      gender = 'Male';
-    }
-    return gender;
-  };
-
-  const addCategory = () => {
-    if (checkCategoryFormErrors() > 1) {
+    if (checkForGeneralInfoFormErrors() > 1) {
+      alert('Please fill in all the required fields');
       return;
     }
 
-    if (!checkIfNameIsUnique(categories, categoryName)) {
-      alert('Category name must be unique');
-      return;
-    }
-
-    setCategories([
-      {
-        name: categoryName,
-        gender: getGender(),
-        minAge,
-        maxAge,
-        minWeight,
-        maxWeight,
-      },
-      ...categories,
-    ]);
-    resetCategoryForm();
-  };
-
-  const editCategory = () => {
-    if (checkCategoryFormErrors() > 1) {
-      return;
-    }
-
-    if (!checkIfNameIsUnique(categories, categoryName)) {
-      alert('Category name must be unique');
-      return;
-    }
-
-    const newCategories: Category[] = categories.map((category, index) => {
-      if (index === editedCategoryNumber) {
-        return {
-          name: categoryName,
-          gender: getGender(),
-          minAge,
-          maxAge,
-          minWeight,
-          maxWeight,
-        };
-      }
-      return category;
-    });
-
-    resetCategoryForm();
-    setCategories(newCategories);
-  };
-
-  const handleDelete = (category: Category) => {
-    setCategories(categories.filter((c) => c.name !== category.name));
-  };
-
-  const handleEdit = (category: Category) => {
-    setEditedCategoryNumber(
-      categories.findIndex((c) => c.name === category.name),
-    );
-    setCategoryName(category.name);
-    setMinAge(category.minAge);
-    setMaxAge(category.maxAge);
-    setMinWeight(category.minWeight);
-    setMaxWeight(category.maxWeight);
-  };
-
-  const handleEditCancel = () => {
-    setEditedCategoryNumber(-1);
+    console.log('Add season', categories);
   };
 
   const handleCancel = () => {
@@ -162,6 +37,32 @@ const AddSeason: React.FC = () => {
     if (response) {
       navigate(ROUTES.SEASONS);
     }
+  };
+
+  const handleCategoriesUpdate = (categoryArray: Category[]) => {
+    setCategories(categoryArray);
+  };
+
+  const checkForGeneralInfoFormErrors = () => {
+    let points = 1;
+    if (!name) {
+      points *= errorPointsValues.name;
+    }
+
+    if (!startDate) {
+      points *= errorPointsValues.startDate;
+    }
+
+    if (!endDate) {
+      points *= errorPointsValues.endDate;
+    }
+
+    if (startDate > endDate) {
+      points *= errorPointsValues.startEndRange;
+    }
+
+    setErrorPoints(points);
+    return points;
   };
 
   return (
@@ -179,7 +80,11 @@ const AddSeason: React.FC = () => {
             label="Name"
             onChange={(e) => setName(e.target.value)}
             value={name}
-            errorMessage="You need to provide a name"
+            errorMessage={
+              errorPoints % errorPointsValues.name === 0
+                ? 'You need to provide a name'
+                : null
+            }
           />
 
           <TextField
@@ -187,7 +92,11 @@ const AddSeason: React.FC = () => {
             onChange={(e) => setStartDate(e.target.value)}
             value={startDate}
             type="date"
-            errorMessage="You need to provide a start date"
+            errorMessage={
+              errorPoints % errorPointsValues.startDate === 0
+                ? 'You need to provide a start date'
+                : null
+            }
           />
 
           <TextField
@@ -195,79 +104,17 @@ const AddSeason: React.FC = () => {
             onChange={(e) => setEndDate(e.target.value)}
             value={endDate}
             type="date"
-            errorMessage="You need to provide a end date"
+            errorMessage={
+              errorPoints % errorPointsValues.endDate === 0
+                ? 'You need to provide a end date'
+                : errorPoints % errorPointsValues.startEndRange === 0
+                  ? 'End date must be after start date'
+                  : null
+            }
           />
         </div>
 
-        <div className="tile categories categoriesLayout">
-          <div style={{ flex: 1 }}>
-            <p className="subtitle mb30">Categories</p>
-            <TextField
-              label="Category name"
-              onChange={(e) => setCategoryName(e.target.value)}
-              value={categoryName}
-              errorMessage={
-                errorPoints % errorPointsValues.categoryName === 0
-                  ? 'You need to provide a category name'
-                  : undefined
-              }
-            />
-            <div className="gender">
-              <p className="heading">Gender:</p>
-              <div className="checkboxes">
-                <Checkbox ref={femaleCheckboxRef} label="Female" />
-                <Checkbox ref={maleCheckboxRef} label="Male" />
-              </div>
-            </div>
-
-            <div className="age">
-              <p className="heading">Age:</p>
-              <MinMaxField
-                minValue={minAge}
-                maxValue={maxAge}
-                onMinChange={setMinAge}
-                onMaxChange={setMaxAge}
-                errorMessage={
-                  errorPoints % errorPointsValues.age === 0
-                    ? 'You need to provide valid min and max age range'
-                    : undefined
-                }
-              />
-            </div>
-
-            <div className="weight">
-              <p className="heading">Weight:</p>
-              <MinMaxField
-                minValue={minWeight}
-                maxValue={maxWeight}
-                onMinChange={setMinWeight}
-                onMaxChange={setMaxWeight}
-                errorMessage={
-                  errorPoints % errorPointsValues.weight === 0
-                    ? 'You need to provide valid min and max weight range'
-                    : undefined
-                }
-              />
-            </div>
-
-            <Button
-              value={isEdited ? 'Save' : 'Add category'}
-              onClick={isEdited ? editCategory : addCategory}
-              style={{
-                width: '100%',
-              }}
-            />
-          </div>
-
-          <div style={{ flex: 2 }}>
-            <CategoryTable
-              categories={categories}
-              onDelete={handleDelete}
-              onEdit={handleEdit}
-              onEditCancel={handleEditCancel}
-            />
-          </div>
-        </div>
+        <CategoryForm onUpdate={handleCategoriesUpdate} />
       </div>
 
       <Button onClick={addSeason} value="Add season" />
